don't give output error on not children unless there is really an error and not just no children

+++ src/oWebSocketServer.php
@@ -209,18 +209,20 @@
 					}
 					$this->debug("%s","\nProcess " . $exited_pid . " killed, number left: ".count($this->socketNumbers)."\n","YellowBold");
 				} else if ($exited_pid == -1){
-					$this->debug("%s","\nError on child and was not able to kill child (".count($this->socketNumbers).")\n","YellowBold");
-					if(pcntl_wifexited($status)){
-						$this->debug("%s","\nProcess " . $exited_pid . " exited normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
-					}
+					if( $this->socketNumbers > 0 ){
+						$this->debug("%s","\nError on child and was not able to kill child (".count($this->socketNumbers).")\n","YellowBold");
+						if(pcntl_wifexited($status)){
+							$this->debug("%s","\nProcess " . $exited_pid . " exited normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
+						}
 
-					if(pcntl_wifstopped($status)){
-						$this->debug("%s","\nProcess " . $exited_pid . " is stopped but did not exit normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
-						$this->stoppedProcesses[] = $exited_pid;
-					}
+						if(pcntl_wifstopped($status)){
+							$this->debug("%s","\nProcess " . $exited_pid . " is stopped but did not exit normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
+							$this->stoppedProcesses[] = $exited_pid;
+						}
 
-					if(pcntl_wifsignaled($status)){
-						$this->debug("%s","\nProcess " . $exited_pid . " was termined from an uncaught signal and did not exit normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
+						if(pcntl_wifsignaled($status)){
+							$this->debug("%s","\nProcess " . $exited_pid . " was termined from an uncaught signal and did not exit normally, remaining: ".count($this->socketNumbers)."\n","YellowBold");
+						}
 					}
 				}
